# perf(graphics): Implement sorted draw calls