--- a/src/engine/Graphics/Context/draw-call.ts
+++ b/src/engine/Graphics/Context/draw-call.ts
@@ -1,11 +1,14 @@
-import { Matrix } from '../../excalibur';
+import { Matrix } from '../../Math';
 import { ExcaliburGraphicsContextState } from './ExcaliburGraphicsContext';
 
 export class DrawCall {
   public z: number = 0;
   public priority: number = 0;
   public renderer: string;
-  public transform: Matrix;
-  public state: ExcaliburGraphicsContextState;
+  public transform: Matrix = Matrix.identity();
+  public state: ExcaliburGraphicsContextState = {
+    z: 0,
+    opacity: 1
+  };
   public args: any[];
 }