===:clipboard: PR Checklist :clipboard:===

- [ ] :pushpin: issue exists in github for these changes
- [x] :microscope: existing tests still pass
- [x] :see_no_evil: code conforms to the [style guide](https://github.com/excaliburjs/Excalibur/blob/main/STYLEGUIDE.md)
- [x] :triangular_ruler: new tests written and passing / old tests updated with new scenario(s)
- [x] :page_facing_up: changelog entry added (or not needed)

==================

This PR improves graphics performance when switching renderer plugins back and forth. Before this PR it was possible to gridlock drawing during a sufficiently complex debug draw.

This captures the current draw calls and sorts them by z-index, priority, then renderer name break ties.

Benchmark 2000 Actor 59 -> 58 FPS (Relatively small decrease, graph roughly the same)

![image](https://user-images.githubusercontent.com/612071/162594523-9f31a42c-fb8a-4360-a6a3-0a4e1daa6561.png)

Benchmark 4000 Actor 37 -> 40 FPS graph mostly better

![image](https://user-images.githubusercontent.com/612071/162594505-c3d01a32-8101-41b9-8448-9dd92fd6c2d0.png)